+++ src/bundle.rs
@@ -6,13 +6,16 @@ use memuse::DynamicUsage;
 use redjubjub::{Binding, SpendAuth};
 
 use zcash_note_encryption::{
-    EphemeralKeyBytes, ShieldedOutput, COMPACT_NOTE_SIZE, ENC_CIPHERTEXT_SIZE, OUT_CIPHERTEXT_SIZE,
+    note_bytes::NoteBytesData, EphemeralKeyBytes, ShieldedOutput, OUT_CIPHERTEXT_SIZE,
 };
 
 use crate::{
     constants::GROTH_PROOF_SIZE,
     note::ExtractedNoteCommitment,
-    note_encryption::{CompactOutputDescription, SaplingDomain},
+    note_encryption::{
+        CompactOutputDescription, NoteCiphertextBytes, SaplingDomain, COMPACT_NOTE_SIZE,
+        ENC_CIPHERTEXT_SIZE,
+    },
     value::ValueCommitment,
     Nullifier,
 };
@@ -334,7 +337,7 @@ pub struct OutputDescription<Proof> {
     cv: ValueCommitment,
     cmu: ExtractedNoteCommitment,
     ephemeral_key: EphemeralKeyBytes,
-    enc_ciphertext: [u8; ENC_CIPHERTEXT_SIZE],
+    enc_ciphertext: NoteCiphertextBytes,
     out_ciphertext: [u8; OUT_CIPHERTEXT_SIZE],
     zkproof: Proof,
 }
@@ -356,7 +359,7 @@ impl<Proof> OutputDescription<Proof> {
 
     /// Returns the encrypted note ciphertext.
     pub fn enc_ciphertext(&self) -> &[u8; ENC_CIPHERTEXT_SIZE] {
-        &self.enc_ciphertext
+        &self.enc_ciphertext.0
     }
 
     /// Returns the output recovery ciphertext.
@@ -382,7 +385,7 @@ impl<Proof> OutputDescription<Proof> {
             cv,
             cmu,
             ephemeral_key,
-            enc_ciphertext,
+            enc_ciphertext: NoteBytesData(enc_ciphertext),
             out_ciphertext,
             zkproof,
         }
@@ -401,7 +404,7 @@ impl<Proof> OutputDescription<Proof> {
         &mut self.ephemeral_key
     }
     pub(crate) fn enc_ciphertext_mut(&mut self) -> &mut [u8; ENC_CIPHERTEXT_SIZE] {
-        &mut self.enc_ciphertext
+        &mut self.enc_ciphertext.0
     }
     pub(crate) fn out_ciphertext_mut(&mut self) -> &mut [u8; OUT_CIPHERTEXT_SIZE] {
         &mut self.out_ciphertext
@@ -418,17 +421,25 @@ impl<Proof: DynamicUsage> DynamicUsage for OutputDescription<Proof> {
     }
 }
 
-impl<A> ShieldedOutput<SaplingDomain, ENC_CIPHERTEXT_SIZE> for OutputDescription<A> {
+impl<A> ShieldedOutput<SaplingDomain> for OutputDescription<A> {
     fn ephemeral_key(&self) -> EphemeralKeyBytes {
         self.ephemeral_key.clone()
     }
 
-    fn cmstar_bytes(&self) -> [u8; 32] {
-        self.cmu.to_bytes()
+    fn cmstar(&self) -> &ExtractedNoteCommitment {
+        &self.cmu
     }
 
-    fn enc_ciphertext(&self) -> &[u8; ENC_CIPHERTEXT_SIZE] {
-        &self.enc_ciphertext
+    fn enc_ciphertext(&self) -> Option<&NoteCiphertextBytes> {
+        Some(&self.enc_ciphertext)
+    }
+
+    fn enc_ciphertext_compact(&self) -> NoteBytesData<COMPACT_NOTE_SIZE> {
+        NoteBytesData(
+            self.enc_ciphertext.0[..COMPACT_NOTE_SIZE]
+                .try_into()
+                .expect("slice is the correct length"),
+        )
     }
 }
 
@@ -479,7 +490,7 @@ impl OutputDescriptionV5 {
             cv: self.cv,
             cmu: self.cmu,
             ephemeral_key: self.ephemeral_key,
-            enc_ciphertext: self.enc_ciphertext,
+            enc_ciphertext: NoteBytesData(self.enc_ciphertext),
             out_ciphertext: self.out_ciphertext,
             zkproof,
         }
@@ -491,7 +502,9 @@ impl<A> From<OutputDescription<A>> for CompactOutputDescription {
         CompactOutputDescription {
             ephemeral_key: out.ephemeral_key,
             cmu: out.cmu,
-            enc_ciphertext: out.enc_ciphertext[..COMPACT_NOTE_SIZE].try_into().unwrap(),
+            enc_ciphertext: out.enc_ciphertext.0[..COMPACT_NOTE_SIZE]
+                .try_into()
+                .unwrap(),
         }
     }
 }
@@ -518,7 +531,7 @@ pub mod testing {
     };
 
     use super::{
-        Authorized, Bundle, GrothProofBytes, OutputDescription, SpendDescription,
+        Authorized, Bundle, GrothProofBytes, NoteBytesData, OutputDescription, SpendDescription,
         ENC_CIPHERTEXT_SIZE, OUT_CIPHERTEXT_SIZE,
     };
 
@@ -581,7 +594,7 @@ pub mod testing {
                 cv,
                 cmu,
                 ephemeral_key: epk.to_bytes().into(),
-                enc_ciphertext,
+                enc_ciphertext: NoteBytesData(enc_ciphertext),
                 out_ciphertext,
                 zkproof,
             }