+++ src/pczt/signer.rs
@@ -1,6 +1,6 @@
 use core::fmt;
 
-use rand::{CryptoRng, RngCore};
+use rand::{CryptoRng, Rng};
 use redjubjub::SpendAuth;
 
 use crate::keys::SpendAuthorizingKey;
@@ -11,7 +11,7 @@ impl super::Spend {
     /// It is the caller's responsibility to perform any semantic validity checks on the
     /// PCZT (for example, comfirming that the change amounts are correct) before calling
     /// this method.
-    pub fn sign<R: RngCore + CryptoRng>(
+    pub fn sign<R: Rng + CryptoRng>(
         &mut self,
         sighash: [u8; 32],
         ask: &SpendAuthorizingKey,